Move plan cmd to kubectl plugin
