--- cmd/schemahero/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/schemahero/schemahero/pkg/cli/schemaherocli"
-
-func main() {
-	schemaherocli.InitAndExecute()
-}
